moving all json file to config dir
	modified:   README.md
	renamed:    example-configurations/example-autoop.json -> config/example-autoop.json
	renamed:    example-configurations/example-config.json -> config/example-config.json
	renamed:    example-configurations/example-greetings.json -> config/example-greetings.json
	modified:   modules/autoop.js
	modified:   modules/greetings.js
	modified:   modules/is-admin.js
	modified:   modules/is-channel.js
