+++ README.md
@@ -12,7 +12,7 @@ Features
 - Add new functionality (commands and observers) on the fly without rebooting the bot!
 - Bot can have 1 or many admins
 - Auto-op bot admin(s). Requires the bot to have `+o` of course.
-- Auto-op users defined in `./autoop.json`. **WARNING**: Auto-op is nick based so use cautiously (if at all)!
+- Auto-op users defined in `./config/autoop.json`. **WARNING**: Auto-op is nick based so use cautiously (if at all)!
 - Optional per channel greetings for newly joined user defined in `./greetings.json`
 - Admin can send basic IRC commands like `/mode`, `/join`, `/part` and make the bot say things
