+++ modules/greetings.js
@@ -1,4 +1,4 @@
-var config  = require('../config.json');
+var config  = require('../config/config.json');
 var fs = require('fs');
 /*
  *
@@ -9,8 +9,8 @@ module.exports = function(client, channel, nick, message) {
 
   var greetings;
 
-  if (fs.existsSync('../greetings.json')) {
-    greetings = require('../greetings.json');
+  if (fs.existsSync('../config/greetings.json')) {
+    greetings = require('../config/greetings.json');
 
     // if greetins.json configured, then greet the newly joined user
     if (nick != config.userName, greetings != undefined && typeof greetings[channel] == "string" && greetings[channel].length > 1) {
